requirements.txt
anyio==4.4.0
async-property==0.2.2
black==24.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
deprecation==2.1.0
et_xmlfile==2.0.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
jwcrypto==1.5.6
mypy==1.13.0
mypy-extensions==1.0.0
numpy==2.1.3
openpyxl==3.1.5
packaging==24.0
pandas==2.2.3
pandas-stubs==2.2.3.241126
pathspec==0.12.1
platformdirs==4.2.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-keycloak==4.2.0
pytz==2024.2
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
tomli==2.0.1
types-pytz==2024.2.0.20241003
types-requests==2.32.0.20241016
typing_extensions==4.11.0
tzdata==2024.2
urllib3==2.2.1
